+++ Cargo.toml
@@ -16,13 +16,13 @@ paw-attributes = { path = "paw-attributes", version = "0.0.0" }
 paw-raw = { path = "paw-raw", version = "0.0.0" }
 
 [dev-dependencies]
-paw-clap = { path = "paw-clap", version = "0.0.0" }
+paw-structopt = { path = "paw-structopt", version = "0.0.0" }
 structopt = "0.2.15"
 
 [workspace]
 members = [
   ".",
   "paw-attributes",
-  "paw-clap",
+  "paw-structopt",
   "paw-raw",
 ]
